refactor preprocess_and_predict
