+++ src/predictions/profiles_mlcorelib/ml_core/preprocess_and_predict.py
@@ -3,8 +3,6 @@ import json
 import time
 import logging
 import warnings
-import cachetools
-import numpy as np
 import pandas as pd
 from typing import Any, List, Dict, Generator, Tuple
 from tqdm import tqdm
@@ -33,97 +31,15 @@ warnings.simplefilter("ignore", category=NumbaPendingDeprecationWarning)
 DEFAULT_BATCH_SIZE = 10000
 
 
-def get_batch_iterator(
-    connector: Connector,
-    joined_input_table_name: str,
-    batch_size: int,
-    filter_condition: str = None,
-) -> Generator[pd.DataFrame, None, None]:
-    """
-    Creates a generator that yields batches of data from the input table.
-    For BigQuery and Redshift, this uses LIMIT and OFFSET to fetch data in batches.
-    """
-    count_query = f"SELECT COUNT(*) as count FROM {joined_input_table_name}"
-    if filter_condition:
-        count_query += f" WHERE {filter_condition}"
-
-    total_rows = connector.run_query(count_query)[0].count
-    total_batches = int(np.ceil(total_rows / batch_size))
-    logger.get().info(
-        f"Starting batch processing with {total_batches} batches (batch_size={batch_size})"
-    )
-
-    def iterator():
-        for offset in range(0, total_rows, batch_size):
-            query = f"SELECT * FROM {joined_input_table_name}"
-            if filter_condition:
-                query += f" WHERE {filter_condition}"
-            query += f" LIMIT {batch_size} OFFSET {offset}"
-
-            batch_df = connector.run_query(query)
-            yield pd.DataFrame(batch_df)
-
-    return iterator(), total_batches
-
-
-def process_batch(
-    batch_df: pd.DataFrame,
-    numeric_columns: List[str],
-    categorical_columns: List[str],
-    arraytype_columns: List[str],
-    booleantype_columns: List[str],
-    timestamp_columns: List[str],
-    ignore_features: List[str],
-    required_features_upper_case: set,
-    transformed_arraytype_columns: Dict[str, List[str]],
-    trainer: MLTrainer,
-    connector: Connector,
-    end_ts: str,
-) -> Tuple[pd.DataFrame, pd.DataFrame]:
-    """Process a single batch of data for preprocessing."""
-    logger.get().debug(f"Processing batch of size {len(batch_df)}")
-
-    _, batch_df = connector.transform_arraytype_features(
-        batch_df,
-        arraytype_columns,
-        trainer.prep.top_k_array_categories,
-        predict_arraytype_features=transformed_arraytype_columns,
-    )
-
-    batch_df = connector.transform_booleantype_features(batch_df, booleantype_columns)
-
-    input_df = connector.drop_cols(batch_df, ignore_features)
-    input_df = connector.select_relevant_columns(input_df, required_features_upper_case)
-
-    predict_data = batch_df[[trainer.entity_column]]
-    predict_data = connector.add_index_timestamp_colum_for_predict_data(
-        predict_data, trainer.index_timestamp, end_ts
-    )
-
-    return predict_data, input_df
-
-
-def preprocess_and_predict(
-    creds,
-    s3_config,
-    model_path,
-    inputs,
-    end_ts,
-    output_tablename,
-    connector: Connector,
-    trainer: MLTrainer,
-    model_hash: str,
-    batch_size: int = DEFAULT_BATCH_SIZE,
+def setup_prediction_environment(
+    model_path: str, connector: Connector, trainer: MLTrainer
 ):
-    """
-    This function is responsible for preprocessing and predicting on the data.
-    For BigQuery and Redshift, it processes data in batches from the start
-    to handle large datasets efficiently.
-    """
+    """Sets up the prediction environment by loading model info and configuring necessary parameters."""
     connector.compute_udf_name(model_path)
 
     with open(model_path, "r") as f:
         results = json.load(f)
+
     train_model_id = results["model_info"]["model_id"]
     stage_name = results["model_info"]["file_location"]["stage"]
     pkl_model_file_name = results["model_info"]["file_location"]["file_name"]
@@ -156,117 +72,79 @@ def preprocess_and_predict(
     whtService = PyNativeWHT(None, None, None)
     whtService.set_connector(connector)
 
+    return {
+        "train_model_id": train_model_id,
+        "stage_name": stage_name,
+        "pkl_model_file_name": pkl_model_file_name,
+        "score_column": score_column,
+        "percentile_column": percentile_column,
+        "output_column": output_column,
+        "model_id_column": model_id_column,
+        "numeric_columns": numeric_columns,
+        "categorical_columns": categorical_columns,
+        "arraytype_columns": arraytype_columns,
+        "timestamp_columns": timestamp_columns,
+        "booleantype_columns": booleantype_columns,
+        "ignore_features": ignore_features,
+        "required_features_upper_case": required_features_upper_case,
+        "input_columns": input_columns,
+        "transformed_arraytype_columns": transformed_arraytype_columns,
+        "end_ts": end_ts,
+        "whtService": whtService,
+    }
+
+
+def setup_joined_input_table(
+    connector: Connector,
+    inputs: List[utils.InputsConfig],
+    input_columns: List[str],
+    trainer: MLTrainer,
+) -> str:
+    """Creates a joined input table from the input configurations."""
     joined_input_table_name = (
         f"prediction_joined_table_{utils.generate_random_string(5)}"
     )
     connector.join_input_tables(
         inputs, input_columns, trainer.entity_column, joined_input_table_name
     )
+    return joined_input_table_name
 
-    @cachetools.cached(cache={})
-    def load_model(filename: str):
-        """session.import adds the staged model file to an import directory. We load the model file from this location"""
-        import_dir = sys._xoptions.get("snowflake_import_directory")
 
-        if import_dir:
-            assert import_dir.startswith("/home/udf/")
-            filename = os.path.join(import_dir, filename)
-        else:
-            filename = os.path.join(local_folder, filename)
-
-        model = trainer.load_model(filename)
-        return model
-
-    def predict_helper(df, pkl_model_file_name: str) -> Any:
-        df.columns = features
-        df = utils.transform_null(
-            df, numeric_columns, categorical_columns, timestamp_columns
-        )
-
-        trained_model = load_model(pkl_model_file_name)
-        df.columns = [x.upper() for x in df.columns]
-        return trainer.predict(trained_model, df)
+def preprocess_and_predict(
+    creds,
+    s3_config,
+    model_path,
+    inputs,
+    end_ts,
+    output_tablename,
+    connector: Connector,
+    trainer: MLTrainer,
+    model_hash: str,
+    batch_size: int = DEFAULT_BATCH_SIZE,
+):
+    """Main function to preprocess data and generate predictions."""
+    env_setup = setup_prediction_environment(model_path, connector, trainer)
+    joined_input_table_name = setup_joined_input_table(
+        connector, inputs, env_setup["input_columns"], trainer
+    )
 
     if creds["type"] == "snowflake":
         logger.get().debug("Using Snowflake native processing")
-        raw_data = connector.get_table(
-            joined_input_table_name, filter_condition=trainer.eligible_users
+        prediction_udf, predict_data, input_df = (
+            connector.execute_snowflake_predictions(
+                trainer, joined_input_table_name, env_setup, end_ts
+            )
         )
-        _, raw_data = connector.transform_arraytype_features(
-            raw_data,
-            arraytype_columns,
-            trainer.prep.top_k_array_categories,
-            predict_arraytype_features=transformed_arraytype_columns,
-        )
-        raw_data = connector.transform_booleantype_features(
-            raw_data, booleantype_columns
-        )
-        input_df = connector.drop_cols(raw_data, ignore_features)
-        input_df = connector.select_relevant_columns(
-            input_df, required_features_upper_case
-        )
-        types = connector.generate_type_hint(input_df)
 
-        predict_data = raw_data.select(trainer.entity_column)
-        predict_data = connector.add_index_timestamp_colum_for_predict_data(
-            predict_data, trainer.index_timestamp, end_ts
-        )
-        features = input_df.columns
-
-        # Snowflake UDF setup
-        udf_name = connector.udf_name
-
-        pycaret_score_column = trainer.pred_output_df_columns["score"]
-        pycaret_label_column = trainer.pred_output_df_columns.get(
-            "label", "prediction_score"
-        )  # To make up for the missing column in case of regression
-
-        @F.pandas_udtf(
-            session=connector.session,
-            is_permanent=False,
-            replace=True,
-            stage_location=stage_name,
-            name=udf_name,
-            output_schema=PandasDataFrameType(
-                [FloatType(), FloatType()], [pycaret_score_column, pycaret_label_column]
-            ),
-            input_types=[PandasDataFrameType(types)],
-            input_names=features,
-            imports=[f"{stage_name}/{pkl_model_file_name}.pkl"]
-            + connector.delete_files,
-            packages=constants.SNOWFLAKE_TRAINING_PACKAGES + ["cachetools==4.2.2"],
-        )
-        class predict_scores:
-            def end_partition(self, df):
-                predictions = predict_helper(df, pkl_model_file_name)
-
-                # Create a new DataFrame with the extracted column names
-                prediction_df = pd.DataFrame()
-                prediction_df[pycaret_score_column] = predictions[pycaret_score_column]
-
-                # Check if 'label' is present in pred_output_df_columns
-                # Had to add a dummy label column in case of regression to the output dataframe as the UDTF expects the two columns in output
-                if "label" in trainer.pred_output_df_columns:
-                    prediction_df[pycaret_label_column] = predictions[
-                        pycaret_label_column
-                    ]
-                else:
-                    prediction_df[pycaret_label_column] = np.nan
-
-                yield prediction_df
-
-        prediction_udf = predict_scores
-
-        # Call prediction UDF
         preds_with_percentile = connector.call_prediction_udf(
             predict_data,
             prediction_udf,
             trainer.entity_column,
             trainer.index_timestamp,
-            score_column,
-            percentile_column,
-            output_column,
-            train_model_id,
+            env_setup["score_column"],
+            env_setup["percentile_column"],
+            env_setup["output_column"],
+            env_setup["train_model_id"],
             input_df,
             trainer.pred_output_df_columns,
         )
@@ -283,141 +161,16 @@ def preprocess_and_predict(
 
     elif creds["type"] in ("redshift", "bigquery"):
         logger.get().debug("Using batch processing for BigQuery/Redshift")
-        local_folder = connector.get_local_dir()
-
-        def predict_scores_rs(df: pd.DataFrame) -> pd.DataFrame:
-            predictions = predict_helper(df, pkl_model_file_name)
-            return predictions
-
-        prediction_udf = predict_scores_rs
-        features = None  # Will be set in first batch
-        first_batch = True
-
-        batch_iterator, total_batches = get_batch_iterator(
-            connector,
-            joined_input_table_name,
-            batch_size,
-            filter_condition=trainer.eligible_users,
+        connector.execute_batch_predictions(
+            trainer, env_setup, output_tablename, s3_config, batch_size
         )
-
-        log_frequency = max(1, total_batches // 10)  # Log every 10% of batches
-        current_batch = 0
-        for batch_df in tqdm(batch_iterator, desc="Processing batches"):
-            current_batch += 1
-            if current_batch % log_frequency == 0:
-                progress_percentage = (current_batch / total_batches) * 100
-                logger.get().info(
-                    f"Processing batch {current_batch}/{total_batches} ({progress_percentage:.1f}% complete)"
-                )
-
-            max_retries = 3
-            retry_count = 0
-            while retry_count < max_retries:
-                try:
-                    # Process batch
-                    batch_predict_data, batch_input_df = process_batch(
-                        batch_df,
-                        numeric_columns,
-                        categorical_columns,
-                        arraytype_columns,
-                        booleantype_columns,
-                        timestamp_columns,
-                        ignore_features,
-                        required_features_upper_case,
-                        transformed_arraytype_columns,
-                        trainer,
-                        connector,
-                        end_ts,
-                    )
-
-                    if features is None:
-                        features = batch_input_df.columns
-
-                    batch_predictions = predict_scores_rs(batch_input_df)
-
-                    batch_result = pd.DataFrame()
-                    batch_result[trainer.entity_column] = batch_predict_data[
-                        trainer.entity_column
-                    ]
-                    batch_result[trainer.index_timestamp] = batch_predict_data[
-                        trainer.index_timestamp
-                    ]
-                    batch_result[model_id_column] = train_model_id
-                    batch_result[score_column] = batch_predictions[
-                        trainer.pred_output_df_columns["score"]
-                    ].round(5)
-                    if "label" in trainer.pred_output_df_columns:
-                        batch_result[output_column] = batch_predictions[
-                            trainer.pred_output_df_columns["label"]
-                        ]
-
-                    logger.get().debug("Writing predictions to warehouse")
-
-                    if first_batch:
-                        connector.write_table(
-                            batch_result,
-                            output_tablename,
-                            write_mode="overwrite",
-                            local=False,
-                            if_exists="replace",
-                            s3_config=s3_config,
-                        )
-                        first_batch = False
-                    else:
-                        values_list = []
-                        for _, row in batch_result.iterrows():
-                            row_values = f"('{row[trainer.entity_column]}', '{row[trainer.index_timestamp]}', '{row[model_id_column]}', {row[score_column]}, {row[output_column]})"
-                            values_list.append(row_values)
-
-                        values_clause = ",\n    ".join(values_list)
-
-                        insert_query = f"""INSERT INTO {output_tablename} ({trainer.entity_column}, {trainer.index_timestamp}, {model_id_column}, {score_column}, {output_column})
-                        VALUES 
-                            {values_clause}"""
-                        connector.run_query(insert_query, response=False)
-
-                    del (
-                        batch_df,
-                        batch_predict_data,
-                        batch_input_df,
-                        batch_predictions,
-                        batch_result,
-                    )
-                    break
-                except Exception as e:
-                    if retry_count < max_retries:
-                        retry_count += 1
-                        time.sleep(5)
-                        continue
-                    else:
-                        drop_table_query = f"DROP TABLE IF EXISTS {output_tablename}"
-                        connector.run_query(drop_table_query, response=False)
-
-                        logger.get().error(f"Error while processing batch: {e}")
-                        raise Exception(f"Error while processing batch: {e}")
-
-        # Calculate percentiles using SQL directly in the warehouse
-        percentile_column_type = "FLOAT" if creds["type"] == "redshift" else "FLOAT64"
-
-        percentile_column_create_query = f"""
-        ALTER TABLE {output_tablename} 
-        ADD COLUMN {percentile_column} {percentile_column_type}
-        """
-
-        percentile_query = f"""
-        UPDATE {output_tablename} t1
-        SET {percentile_column} = ROUND(t2.percentile_rank * 100, 5)
-        FROM (
-            SELECT 
-                {trainer.entity_column},
-                PERCENT_RANK() OVER (ORDER BY {score_column}) as percentile_rank
-            FROM {output_tablename}
-        ) t2
-        WHERE t1.{trainer.entity_column} = t2.{trainer.entity_column}
-        """
-
-        for query in (percentile_column_create_query, percentile_query):
-            connector.run_query(query, response=False)
+        connector.calculate_percentiles(
+            output_tablename,
+            trainer,
+            env_setup["score_column"],
+            env_setup["percentile_column"],
+            creds["type"],
+        )
 
         logger.get().info("Batch processing completed successfully")
 
@@ -427,7 +180,7 @@ def preprocess_and_predict(
             str(end_ts.date()),
             trainer.output_profiles_ml_model,
             model_hash,
-            whtService.get_registry_table_name(),
+            env_setup["whtService"].get_registry_table_name(),
         )
 
         logger.get().debug(
@@ -459,7 +212,7 @@ def preprocess_and_predict(
         metrics_df = pd.DataFrame(
             {
                 "model_name": [trainer.output_profiles_ml_model],
-                model_id_column: [model_id],
+                env_setup["model_id_column"]: [model_id],
                 "prediction_date": [prediction_date],
                 "label_date": [end_ts],
                 "prediction_table_name": [prev_prediction_table],
